+++ src/compas_xr/realtime_database/__init__.py
@@ -18,6 +18,6 @@ Classes
 
 """
 
-from compas_xr.realtime_database.realtime_database_pyrebase import RealtimeDatabase
+from compas_xr.realtime_database.realtime_database import RealtimeDatabase
 
 __all__ = ["RealtimeDatabase"]
